Move handlers to utils
