+++ src/utils/handlers/authenticationHandlers.js
@@ -1,8 +1,6 @@
 import axios from "axios";
-import { pathObj } from "./../utils/generalVariables";
-import { setHeaders } from "./../utils/requestHeaders";
-import { tokenKey } from './../utils/generalVariables';
-
+import { setHeaders } from '../../utils/requestHeaders';
+import { pathObj, tokenKey } from '../../utils/generalVariables';
 
 export const loginHandler = ({ username, password }) => {
   // With the finalization of the database schema, more checks can be implemented (with separate error-messages)
